Use 32bit counters to avoid overflow
    - Previously there were cases where we could overflow cpu65_cycles_to_execute when adding the cycles_speaker_feedback

+++ src/cpu.h
@@ -57,7 +57,7 @@ extern void *cpu65_vmem_w[65536];
 extern unsigned char cpu65_flags_encode[256];
 extern unsigned char cpu65_flags_decode[256];
 
-extern int16_t cpu65_cycle_count;
+extern int32_t cpu65_cycle_count;
 
 #if CPU_TRACING
 void cpu65_trace_begin(const char *trace_file);
